requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
multidict==6.0.4
PyMySQL==1.0.2
yarl==1.8.2
